requirements.txt
newsapi-python==0.2.7
Pillow==10.2.0
requests==2.31.0 